modularize utils
