+++ src/cli/utils/mod.rs
@@ -1,2 +1,14 @@
-pub mod utils;
-pub use utils::*;
+pub mod fs;
+pub use fs::*;
+pub mod process;
+pub use process::*;
+pub mod color;
+pub use color::*;
+pub mod user;
+pub use user::*;
+pub mod check_version;
+pub use check_version::*;
+pub mod dialog;
+pub use dialog::*;
+pub mod update;
+pub use update::*;
