+++ src/cli/utils/user.rs
@@ -1,0 +1,16 @@
+use super::process::async_command_pipe;
+use anyhow::Result;
+use sudo::{check, RunningAs};
+
+pub fn check_root() -> Result<bool> {
+    if let RunningAs::Root = check() {
+        Ok(true)
+    } else {
+        Ok(false)
+    }
+}
+
+pub async fn check_user() -> Result<String> {
+    let user = async_command_pipe("echo ${SUDO_USER:-$USER}").await?;
+    Ok(user)
+}
